+++ CHANGELOG.md
@@ -35,6 +35,8 @@ IMPROVEMENTS:
   - providers/docker: `stop_timeout` can be used to modify the `docker stop`
       timeout. [GH-4504]
   - provisioners/chef: Automatically install Chef when using a Chef provisioner.
+  - provisioners/ansible: Always show Ansible command executed when Vagrant log
+      level is debug (even if ansible.verbose is false)
   - synced\_folders/nfs: Won't use `sudo` to write to /etc/exports if there
       are write privileges. [GH-2643]
   - synced\_folders/smb: Credentials from one SMB will be copied to the rest. [GH-4675]